Move Utils into *.common package
