+++ src/main/java/io/singularitynet/sdk/registry/EndpointGroup.java
@@ -8,6 +8,8 @@ import java.util.ArrayList;
 import com.google.gson.annotations.SerializedName;
 import static com.google.common.base.Preconditions.checkArgument;
 
+import io.singularitynet.sdk.common.Utils;
+
 @EqualsAndHashCode
 @ToString
 public class EndpointGroup {
